refactor(Name): Name의 원시값 포장
이름에 관한 검증 로직은 Name으로 이동시키고 Participant가 Name을 가지도록 함
